Integrating RPMB FS with ENC FS. This was done by redefining tee_fs_common_XX APIs in tee_rpmb_fs_common.c. Based on the compile time flag CFG_RPMB_FS either tee_fs_common.c is chosen or tee_rpmb_fs_common.c.

+++ core/tee/tee_rpmb_fs.c
@@ -1096,7 +1096,7 @@ TEE_Result tee_rpmb_fs_stat(const char *filename,
 	if (res != TEE_SUCCESS)
 		goto out;
 
-	stat->size = fh->fat_entry.data_size;
+	stat->size = (size_t)fh->fat_entry.data_size;
 	stat->reserved = 0;
 
 out:
